requirements.txt
autopep8==2.0.4
djangorestframework==3.14.0
drf-yasg==1.21.7
Django==5.0.2
faker==13.15.1
psycopg2-binary==2.9.9
# oauthlib==2.0.2
